An 3/Sem1/Affective Computing/darknet_model/requirements.txt
torch
torchvision
easydict
Pillow
opencv-python
numpy
